fix: compute camera frame buffer sizes in size_t to avoid int overflow
The frame buffer size calculations in the libVNC and libVLC camera
backends multiplied 16/32-bit operands and only widened the result to
size_t afterwards, so the multiplication itself could overflow before
the conversion. Flagged by CodeQL cpp/integer-multiplication-cast-to-long
(alerts 282, 145, 85). The VNC framebuffer dimensions are advertised by
the remote server (uint16_t), making the inputs externally influenced.

Cast the first operand to size_t so each product is computed in size_t:
- zm_libvnc_camera.cpp: SwScale::Convert src size (framebufferWidth *
  framebufferHeight * 4) and dest size (width * height * colours)
- zm_libvnc_camera.cpp: resize() bufferSize, changed from int to size_t
  (not converted to a larger type, so CodeQL did not flag it, but the
  same overflow applied); Debug format updated to %zu
- zm_libvlc_camera.cpp: Image::Assign size (width * height * mBpp)

Verified: both translation units compile with -Werror -fsyntax-only.

(cherry picked from commit 286cb90883bda53296b44b063b4a1bf4efdeb837)

+++ src/zm_libvlc_camera.cpp
@@ -304,7 +304,7 @@ int LibvlcCamera::Capture(std::shared_ptr<ZMPacket> &zm_packet) {
     return 0;
 
   mLibvlcData.mutex.lock();
-  zm_packet->image->Assign(width, height, colours, subpixelorder, mLibvlcData.buffer, width * height * mBpp);
+  zm_packet->image->Assign(width, height, colours, subpixelorder, mLibvlcData.buffer, static_cast<size_t>(width) * height * mBpp);
   zm_packet->packet->stream_index = mVideoStreamId;
   zm_packet->stream = mVideoStream;
   mLibvlcData.mutex.unlock();
